Removed Permissible
Previously, the permission table had only one identifier for the resource: the resource's ID. Permissible was created to ensure that "permissible" resources wouldn't have conflicting IDs.

The permission table was later updated to support giving permission to a type of resource, or entity set. With the new entity set column, it was no longer necessary to prevent conflicting IDs, and thus permissible is not necessary.